+++ services/repository/repository.go
@@ -28,36 +28,6 @@ func CreateRepository(doer, owner *models.User, opts models.CreateRepoOptions) (
 	return repo, nil
 }
 
-// AdoptRepository adopts pre-existing repository files for the user/organization.
-func AdoptRepository(doer, owner *models.User, opts models.CreateRepoOptions) (*models.Repository, error) {
-	repo, err := repo_module.AdoptRepository(doer, owner, opts)
-	if err != nil {
-		// No need to rollback here we should do this in AdoptRepository...
-		return nil, err
-	}
-
-	notification.NotifyCreateRepository(doer, owner, repo)
-
-	return repo, nil
-}
-
-// DeleteUnadoptedRepository adopts pre-existing repository files for the user/organization.
-func DeleteUnadoptedRepository(doer, owner *models.User, name string) error {
-	return repo_module.DeleteUnadoptedRepository(doer, owner, name)
-}
-
-// ForkRepository forks a repository
-func ForkRepository(doer, u *models.User, opts models.ForkRepoOptions) (*models.Repository, error) {
-	repo, err := repo_module.ForkRepository(doer, u, opts)
-	if err != nil {
-		return nil, err
-	}
-
-	notification.NotifyForkRepository(doer, opts.BaseRepo, repo)
-
-	return repo, nil
-}
-
 // DeleteRepository deletes a repository for a user or organization.
 func DeleteRepository(doer *models.User, repo *models.Repository) error {
 	if err := pull_service.CloseRepoBranchesPulls(doer, repo); err != nil {
